services/notificaciones/requirements.txt
Flask==3.0.0
PyMySQL==1.1.1
python-dotenv==1.0.1
aws-wsgi==0.2.7
boto3==1.34.0